+++ app/Http/Controllers/ImageController.php
@@ -22,6 +22,22 @@ use Kami\Cocktail\Http\Requests\ImageUpdateRequest;
 
 class ImageController extends Controller
 {
+    #[OAT\Get(path: '/images', tags: ['Images'], summary: 'List uploaded images', description: 'List all images uploaded by the authenticated user', parameters: [
+        new BAO\Parameters\PageParameter(),
+        new BAO\Parameters\PerPageParameter(),
+    ])]
+    #[OAT\Response(response: 200, description: 'Successful response', content: [
+        new BAO\PaginateData(BAO\Schemas\Image::class),
+    ])]
+    #[BAO\NotAuthorizedResponse]
+    #[BAO\NotFoundResponse]
+    public function index(Request $request): JsonResource
+    {
+        $images = Image::where('created_user_id', $request->user()->id)->orderBy('created_at', 'desc')->paginate($request->get('per_page', 100));
+
+        return ImageResource::collection($images->withQueryString());
+    }
+
     #[OAT\Get(path: '/images/{id}', tags: ['Images'], summary: 'Show an image', parameters: [
         new BAO\Parameters\DatabaseIdParameter(),
     ])]
@@ -41,10 +57,10 @@ class ImageController extends Controller
         return new ImageResource($image);
     }
 
-    #[OAT\Post(path: '/images', tags: ['Images'], summary: 'Upload an image', requestBody: new OAT\RequestBody(
+    #[OAT\Post(path: '/images', tags: ['Images'], summary: 'Upload an image', description: 'Used to upload multiple images at once. Uploaded images via this endpoint will not be attached to any resource. Images are converted to WebP format with 85% quality of the original image.', requestBody: new OAT\RequestBody(
         required: true,
         content: [
-            new OAT\MediaType(mediaType: 'multipart/form-data', schema: new OAT\Schema(type: 'object', properties: [
+            new OAT\MediaType(mediaType: 'multipart/form-data', schema: new OAT\Schema(type: 'object', required: ['images'], properties: [
                 new OAT\Property(property: 'images', type: 'array', items: new OAT\Items(ref: BAO\Schemas\ImageRequest::class)),
             ])),
         ]
@@ -57,7 +73,7 @@ class ImageController extends Controller
         $manager = ImageManager::imagick();
 
         $images = [];
-        foreach ($request->images as $formImage) {
+        foreach ($request->images ?? [] as $formImage) {
             if (isset($formImage['image'])) {
                 $imageSource = $formImage['image'];
             } else {
@@ -67,14 +83,13 @@ class ImageController extends Controller
             try {
                 $image = new ImageDTO(
                     $manager->read($imageSource),
-                    $formImage['copyright'],
-                    (int) $formImage['sort'],
+                    $formImage['copyright'] ?? null,
+                    (int) ($formImage['sort'] ?? 0),
                 );
+                $images[] = $image;
             } catch (Throwable $e) {
                 Log::error($e->getMessage());
-                abort(500, 'Unable to create an image file!');
             }
-            $images[] = $image;
         }
 
         $images = $imageservice->uploadAndSaveImages($images, $request->user()->id);