+++ routes/api.php
@@ -122,6 +122,7 @@ Route::middleware($apiMiddleware)->group(function () {
     });
 
     Route::prefix('images')->group(function () {
+        Route::get('/', [ImageController::class, 'index']);
         Route::get('/{id}', [ImageController::class, 'show']);
         Route::post('/', [ImageController::class, 'store']);
         Route::post('/{id}', [ImageController::class, 'update']);